requirements.txt
accelerate==0.21.0
appdirs==1.4.4
asttokens==2.2.1
autopep8==2.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
colorama==0.4.6
docker-pycreds==0.4.0
executing==1.2.0
filelock==3.12.2
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
huggingface-hub==0.16.4
icecream==2.1.3
idna==3.4
Jinja2==3.1.2
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.23.4
psutil==5.9.5
pycodestyle==2.11.0
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wandb==0.15.8
